added unit test to test the write to firestore using the security rules in special-parakeet-firestore.rules
