--- Special Parakeet/.babelrc.json
@@ -1,12 +1,0 @@
-{
-  "presets": [["@babel/preset-env", { "targets": { "node": "current" } }]],
-  "plugins": [
-    ["@babel/plugin-transform-runtime",
-      {
-        "regenerator": true
-      }
-    ]
-  ],
-}
-
-
