--- Special Parakeet/babel.config.js
@@ -1,10 +1,0 @@
-module.exports = {
-  presets: [['@babel/preset-env', {targets: {node: 'current'}}]],
-  plugins: [
-    ["@babel/plugin-transform-runtime",
-      {
-        "regenerator": true
-      }
-    ]
-  ],
-};
