--- Special Parakeet/jest.config.js
@@ -1,6 +1,0 @@
-/** @type {import('jest').Config} */
-const config = {
-  verbose: true,
-};
-
-module.exports = config;
